Header and About section
The header has gotten a lot of work and is 95% done, jsut some OCD stuff and bad coding i want to eliminate/repair but thats for the future.

Started working on the about section and trying to implement some transitions when you hover over the "about" section of the site.